Clifford Wolf: 	Tiny 64bit fix
git-svn-id: http://svn.clifford.at/stfl/trunk@127 7279eb3b-6015-0410-83ef-f88855fdc746

+++ parser.c
@@ -448,7 +448,7 @@ parser_error:;
 			fprintf(stderr, " ");
 		else
 		if (*text < 32)
-			fprintf(stderr, "\\%03lo", *text);
+			fprintf(stderr, "\\%03lo", (long unsigned int)*text);
 		else
 			fprintf(stderr, "%lc", (wint_t)*text);
